register remake
